+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 - **Fixed:** Added `wrapperStyle` to `filterDOMProps` defaults. [\#430](https://github.com/vazco/uniforms/issues/430)
+- **Fixed:** `ListAddField` now clones given `value` to not keep its reference. [\#428](https://github.com/vazco/uniforms/issues/428)
 
 ## [v1.24.4](https://github.com/vazco/uniforms/tree/v1.24.4) (2018-05-27)
 - **Fixed:** Placeholders are now working with disabled labels on `GraphQLBridge` and `JSONSchemaBridge`. [\#421](https://github.com/vazco/uniforms/issues/421)